:bug: Fixed artifacts issue
Fixed issue where linking to artifacts in releases leads to 404 error.
The issue occurs because we need to get
job id of the job with artifacts.To do this we get the current
pipeline id list all of the jobs.The user then specifies what jobs
they want artifacts from and we try to match it.